Recognize and record wake-on-wlan in Device.info (#544)
This will allow users of this library to detect whether a Roku device is able to use Wake on Lan magic packets to power on. The application can then choose to send this magic packet to the retained MAC address of this device to wake it, even from deep sleep where it cannot be woken up through keypress/poweron.